Write instructions, try to change up urlconf.
The browser wants to send me to index.html, but I need to redirect it to
index/ instead. I don't seem to be catching the request to just the
website before the browser catches it. Maybe a real issue was being
hidden before and my urlconf doesn't actually work like I think it does.