Start on the login form.
I also need to set up the field in the Floor model to say whether you
can put more stocks in it. I don't know a good way to do that. An
enum would be nice, but alas. Anyway, I need to write up the template
and then set up the view to work right, and then I have to figure out
how I'm going to deal with all the other pages that are necessary with
this (change password, forgot password, etc.) I can't really use the
default ones, but that would be quite easy.