Get floor joining working.
I still need to format the HTML, but the backend works right now. Also,
I figured out what a permanent redirect is. It tells the browser to
cache it, so I shouldn't use it if that page might ever not redirect to
the same place. Never making that mistake again.