Start on custom StockWidget
I got the text box to autofill with typeahead.js. It's really nice, and
it's not even that slow. Kinda. It needs some styling, but that's fine.
I also need to have a box that holds the inputted stocks, but I'll get
to that later. Anyway, things are looking pretty good for this. I'm
excited.

Also, I'm going to have to set up a sorting algorithm because
alphabetical order sucks, and the server is making too many queries, so
that's an issue, and spaces screw everything up. But we're getting
there.