Start to make the stock picker work in python
Right now, it doesn't do anything. I need to override something so
that the python form will see it as valid, but I don't know what.
Something is looking at it before everything I changed runs. Maybe I
need to change the `validators` variable on something. I dunno. I'm
tired, so I'll worry about this later. Also, before I can really do
anyhting, I'll need to change the API setup because the current one will
totally crash if I make a bunch of requests in a row, which I'll need to
do to make this work. So that'll be fun.