Fix StockWidget implementation
It was bad, now it's less bad. Basically, what I did was set it up so
that all you have to do is add the Field in your form, wire up the media
variable in your template, and your good to go. All this ugly stuff the
old docstring made you do is gone.
I mean, it's not gone now. Whenever I get free, it will be. But you
don't have to do it anymore anyhow.
Finish #10.