Try to make the prefetch URL mutable
It's not working because the form can only take the POST request as an
argument. I think. Actually, since it's not being unpacked, maybe not.
Hmm. I'll have to override __init__, but that's OK. So that's what I'll
do. Anyway, it's not working now, but there's just some syntax I have to
clear up, override that __init__ method, and everything should be good.
I think.