Start on trade page
There are a lot of issues yet.
1. The player picking field doesn't do anything.

2. I don't know if the stock picking field will work right with regard
to having something in there at the beginning.

3. The javascript for the stock field is not working. I have to tell the
javascript files what needs them specifically (probably get the id
somehow) so they only act on that thing. It's not terribly efficient,
but it will probably be good for now.

4. The player picking field doesn't do anything and hasn't even been
designed. That will be a problem, one of these days.

5. Also, the stock fields aren't getting their inputs the way they are
supposed to. They are just giving me all the stocks instead of just
some, and they are pinging the general URL instead of the specific one.