Start on the script for the Trade form
I'm having other troubles, so there's nothing in the actual script yet, but someday.
I was trying to give it some media so that it would actually load the
script, but it was removing everything else. I'm sure it's really not
that hard to get it to go, but I haven't figured it out yet. Grep for
TODO to see where the problem is.