Implement validation of Trade page in Django
Make some more progress on #12. I've still got more yet to go, so I
won't be closing it or merging for awhile, but things are moving
forward. There is a comment detailing this, but the code isn't set up
very nicely, so I should probably fix that, but it works. In any case,
see the issue for more info on what's coming next.