Improve admin
Add function to reset floor to initial state in case of buggy trades.
You can get to it from the dropdown thing or whatever it's called. You
know. So that's good. I may have been a little liberal with it, but so
it goes. I should probably add something to make sure that I don't wreck
the whole production database at some point, but that sounds like
something for tomorrow.