Make a lot of progress on adding the trade inbox
I think it's pretty much ready. I want to do some more playing around
before I merge, and maybe do some design (none of this is exactly
pretty, but this is some Dante level terribleness). I was thinking about
finishing the received trade view before I merge so that I can merge
this feature in whole, but then I'd have to do it here or have one
branch not work. I'll just merge it and do it in the other branch. It
doesn't really matter.