Make some progress on fancy trades
I thought that this would do it, but it didn't. It did something though.
Anyway, so I still have to add the code for actually accepting the
suggested stocks, which I forgot about. That shouldn't be so hard
though. Ummm, I also need to add the notifications on the left bar for
the suggestions, and I need to make the Suggested Stocks tab disappear
if the floor isn't permissive.