Fix bug with stock widget on trade page
It didn't know that the floor was an object and not a string, so when it
put it back after you decline to empty the holder box, it would just put
in `[object Object]` instead of "Floor". So I made sure that it only
get's the name all the time. It works just fine now. I should probably
keep checking for bugs though, because I apparently wasn't paying very
close attention when I wrote this and I made a lot of mistakes. Oops.