Finish unit test
It's not half bad. I have a recipe for it in the makefile, and then I
have the test. I found that adding the scores works fine, and is fast
enough that it's not a big deal, I think. Although that's just with 1000
users. With more, things might get complicated. But I think it'll be
fine.