Start on #14
Add view and URL.
Start on template. Add a new default image for the placeholder in the
user page. I still need to add the table for the floors on the page. I
think that I should move the code to generate that table somewhere else
so I DRY, and I also have to change the dashboard code to direct to the
user page instead of the trade page when you click on a user.