Start on #17
I added some fields to the floor model and the the form. I still need
to add some JavaScript warnings in case the user doesn't include enough
stocks, and then I need to make the trade page check that a trade isn't
breaking the rules of the floor before it lets it go through. Also,
maybe I should have a little counter on the trade page that shows how
many stocks you can have and how many you will have after a trade goes
through while you're working on it. That would be slick. Anyway, I'll
get to that stuff.