Work on error checking on received trades
Now it will tell you if you try to accept a trade that can't be
accepted, but it will just show a like a 404 error, not a form error.
I'm working on a way to make it show something more user-friendly. I'm
not 100% sure how to do it. I was thinking I could check when you first
load the received trade page, but that would allow for someone to accept
a trade that has become impossible while he was on the page but before
he clicked "Accept". I think I would be willing to live with that
though. So I'll probably do that. In other news, I put in something to
cause an error so I don't forget to work on this next time.