Fix some little bugs
How did these get past me before? How dumb was I? Anyway, there was a
JavaScript issue with the tabs on the dashboard that I fixed, and there
was an issue with StockSuggestions that would result in you getting the
stock immediately anyway without the person accepting the request. How
foolish.