Improve unit tests
They all pass if there are only 100 users, although it takes an age. If
there are 1000, you start to hit problematic stocks, but it takes so
long that I would rather just leave it for now and get to it later. I
also added a print statement and fixed some serious bugs in the tests.