Make join link work better
When you go to a private floor's edit page, it gives you a link to send
to other people so they can join it. It used to look like a link but
just be a span, but I changed it so it was actually a link, and made it
so that if you go to it when you're already a member of the floor, it
doesn't do anything. Now you can right click and copy the link
instantly. That's really what I was trying to do, although I also
accidentally made the floor joining code more robust.