Fix some bugs from two commits ago
I missed a couple of things, but this fixes them. I made it so that
you can't have more than one floor with a certain name, so that solved a
lot of problems, and then I made errors that tell you that. I also
refactored the view a little.
I think this is good to merge, as long as the tests pass.