Add caching
Now it doesn't have query the server every time you change tabs if you
have downloaded some of those stocks before. I think that it could be a
little better, but it's pretty easy to read and fast as hell, so it's
better than the pity caching that JavaScript was doing for me before.