Fix #40
There was a bug where you could click on stocks on floors that you
aren't on if you go on a User's page, causing a 500 error. I fixed it by
making it such that you can render the stock board without links, which
I then set to render on the User pages. I left the dashboard unchanged.
I think it works, but I need to test it more to make sure that I didn't
miss anything.

Close #40