Add the last endpoints that I forgot
The ones to accept and deny `StockSuggestion`s and `Trade`s weren't
there, and I forgot that I specced them and never wrote them. They
aren't tested yet, but I am on a time crunch and in doing the other
thing that I'm doing I'll test them, so it's not a big deal. Soon,
though.