Fix a bug
There was a problem where if you ask for a single player, when the
trades are being suppressed, I had a naming error. I fixed it. It made
it through the tests because there were no trades in that test, so
everything looked fine. Fun fact: if you mess up your tests, things get
through them. Anyway, I fixed the tests and the code so now that won't
happen again.