Try to fix #74
I made a lot of progress, but now there's a company in the system that
doesn't exist anymore and I don't know how to get rid of it, so the
tests won't pass. We will have to see what can be done about that.
Apparently fixtures are secretly terrible. So that's good to know.